update by 多仔ヾ

+++ admin/adminuserlist.php
@@ -4,6 +4,30 @@ require_once $_SERVER['DOCUMENT_ROOT']."/admin/authorize.php";
 
 $mysqlObj = new sqlhelper();
 
+if(request::post("action") == "delete") {
+    $userId = request::post("id");
+    if(!$userId || $userId == "") {
+        // 未传递用户id
+        response::falure("用户id不能为空");
+    } else if($userId == $_SESSION["adminUser"]["id"]) {
+        // 不能删除自己
+        response::falure("禁止删除");
+    }
+    
+    $sqlStr = "SELECT * FROM sys_admin_user WHERE id = $userId";
+    $result = $mysqlObj->executeQuery($sqlStr);
+    
+    if(count($result) == 0) {
+        // 用户id不存在
+        response::falure("用户id不存在");
+    }
+    
+    $sqlStr = "DELETE FROM sys_admin_user WHERE id = '$userId'";
+    $mysqlObj->executeUpdate($sqlStr);
+    
+    response::success("删除成功");
+}
+
 // 拼接查询条件
 $sqlStr_search = " WHERE 1=1";
 if (request::get("searchAccount") != "") {
@@ -30,7 +54,6 @@ $dataList = $mysqlObj->executeQuery($sqlStr_list);
 <!DOCTYPE html>
 <html lang="cn">
 <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
     <title>管理用户列表 - 数据管理系统(管理端)</title>
     <?php require_once $_SERVER['DOCUMENT_ROOT'] . "/admin/layouts/head.php";?>
 </head>
@@ -65,7 +88,7 @@ $dataList = $mysqlObj->executeQuery($sqlStr_list);
                                                     <input type="text" class="form-control" name="searchAccount" value="<?php echo request::get("searchAccount")?>">
                                                 </div>
                                             </div>
-                                            <div class="col-lg-3 col-12 mb-3s">
+                                            <div class="col-lg-3 col-12 mb-3">
                                                 <div class="input-group">
                                                     <span class="input-group-text">用户备注</span>
                                                     <input type="text" class="form-control" name="searchRemarks" value="<?php echo request::get("searchRemarks")?>">
@@ -77,7 +100,7 @@ $dataList = $mysqlObj->executeQuery($sqlStr_list);
                                         </div>
                                     </form>
                                 </div>
-                                <div class="card-body p-0">
+                                <div class="card-body p-0 table-responsive text-nowrap">
                                     <table class="table table-striped">
                                         <thead>
                                         <tr>
@@ -115,11 +138,16 @@ $dataList = $mysqlObj->executeQuery($sqlStr_list);
                                                     echo "<td><span class='badge text-bg-danger'>禁用</span></td>";
                                                 }
                                                 
-                                                echo "<td>
-                                                    <a class='link-underline link-underline-opacity-0' href='/admin/editadminuser.php?id=".$value["id"]."'><i class='bi bi-pencil-square'></i> 修改</a>
-                                                    ｜
-                                                    <a class='link-underline link-underline-opacity-0' href='javascript:deleteById(".$value["id"].");'><i class='bi bi-trash-fill'></i> 删除</a>
-                                                </td>";
+                                                // 当前用户不能直接编辑自己或删除自己
+                                                if($_SESSION["adminUser"]["id"] == $value["id"]) {
+                                                    echo "<td>-</td>";
+                                                } else {
+                                                    echo "<td>
+                                                        <a class='link-underline link-underline-opacity-0' href='/admin/editadminuser.php?id=".$value["id"]."'><i class='bi bi-pencil-square'></i> 修改</a>
+                                                        ｜
+                                                        <a class='link-underline link-underline-opacity-0' href='javascript:deleteById(".$value["id"].");'><i class='bi bi-trash-fill'></i> 删除</a>
+                                                    </td>";
+                                                }
                                                 echo "</tr>";
                                             }
                                         ?>
@@ -150,8 +178,26 @@ $dataList = $mysqlObj->executeQuery($sqlStr_list);
             $("#searchForm").submit();
         }
         
-        function deleteById(id) {
-        
+        function deleteById(userId) {
+            if (confirm("确定要删除该数据吗？")) {
+                $.ajax({
+                    url: "",
+                    type: "POST",
+                    data: {
+                        id: userId,
+                        action: "delete"
+                    },
+                    dataType: "JSON",
+                    success: function(data) {
+                        if (data.code == 200) {
+                            alert("删除成功");
+                            window.location.reload();
+                        } else {
+                            alert(data.message);
+                        }
+                    }
+                })
+            }
         }
     </script>
 </body>
